requirements.txt
streamlit
pandas
numpy
openpyxl
plotly
streamlit-aggrid
psycopg2-binary